chore: 검색창 custom fetch hook으로 구현
